+++ spec/workbook/worksheet/drawing/shape/shape_properties_spec.rb
@@ -8,6 +8,9 @@ describe 'My behaviour' do
 
   it 'ShapesSpacing' do
     xlsx = OoxmlParser::XlsxParser.parse_xlsx('spec/workbook/worksheet/drawing/shape/properties/shape_spacing.xlsx')
-    expect(xlsx.worksheets[0].drawings[0].shape.text_body.paragraphs[0].properties.spacing).to eq(OoxmlParser::Spacing.new(4, 3, 5, :exact))
+    expect(xlsx.worksheets[0].drawings[0].shape.text_body.paragraphs[0].properties.spacing).to eq(OoxmlParser::Spacing.new(OoxmlParser::OoxmlSize.new(4, :centimeter),
+                                                                                                                           OoxmlParser::OoxmlSize.new(3, :centimeter),
+                                                                                                                           OoxmlParser::OoxmlSize.new(5, :centimeter),
+                                                                                                                           :exact))
   end
 end