+++ spec/workbook/worksheet/drawing/shape/text_body/paragraphs/properties/spacing_spec.rb
@@ -1,0 +1,8 @@
+require 'spec_helper'
+
+describe 'My behaviour' do
+  it 'ShapesRuns' do
+    xlsx = OoxmlParser::Parser.parse('spec/workbook/worksheet/drawing/shape/text_body/paragraphs/properties/spacing/spacing_after_ooxmlsize.xlsx')
+    expect(xlsx.worksheets.first.drawings.first.shape.text_body.paragraphs.first.properties.spacing.after).to eq(OoxmlParser::OoxmlSize.new(2.54, :centimeter))
+  end
+end